+++ packages/compiler-cli/src/ngtsc/metadata/index.ts
@@ -8,7 +8,6 @@
 
 export * from './src/api';
 export {DtsMetadataReader} from './src/dts';
-export {flattenInheritedDirectiveMetadata} from './src/inheritance';
 export {CompoundMetadataRegistry, LocalMetadataRegistry, InjectableClassRegistry} from './src/registry';
 export {extractDirectiveTypeCheckMeta, CompoundMetadataReader} from './src/util';
 export {BindingPropertyName, ClassPropertyMapping, ClassPropertyName, InputOrOutput} from './src/property_mapping';
